requirements.txt
colorama==0.4.6
PyQt5==5.15.9
pyqt5-tools==5.15.9.3.3
python-dotenv==1.0.1
urllib3==2.6.0
requests==2.32.5
yt-dlp==2025.12.8
